pyproject.toml
[project]
name = "workout-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.14.0",
    "bcrypt==4.2.1",
    "email-validator==2.2.0",
    "fastapi[all]==0.115.5",
    "google-cloud-aiplatform>=1.64",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "psycopg2-binary==2.9.10",
    "pydantic[email]==2.10.2",
    "pyjwt==2.10.1",
    "python-dotenv==1.0.1",
    "sqlalchemy[asyncio]==2.0.36",
    "uvicorn==0.32.1",
    "gunicorn>=22.0.0",
    "asyncpg>=0.30.0",
    "supabase>=2.16.0",
    "redis>=6.2.0",
    "ipykernel>=6.29.5",
    "google-cloud-secret-manager>=2.24.0",
]
